Move chart and series utils out of views/alerts
