+++ static/app/chartcuterie/metricChartOption.tsx
@@ -13,8 +13,6 @@ import type {SessionApiResponse} from 'sentry/types/organization';
 import {getCrashFreeRateSeries} from 'sentry/utils/sessions';
 import type {MetricRule, Trigger} from 'sentry/views/alerts/rules/metric/types';
 import {AlertRuleTriggerType, Dataset} from 'sentry/views/alerts/rules/metric/types';
-import {getAnomalyMarkerSeries} from 'sentry/views/alerts/rules/metric/utils/anomalyChart';
-import {isCrashFreeAlert} from 'sentry/views/alerts/rules/metric/utils/isCrashFreeAlert';
 import type {Anomaly, Incident} from 'sentry/views/alerts/types';
 import {IncidentActivityType, IncidentStatus} from 'sentry/views/alerts/types';
 import {
@@ -26,6 +24,8 @@ import {
 } from 'sentry/views/alerts/utils';
 import {AlertWizardAlertNames} from 'sentry/views/alerts/wizard/options';
 import {getAlertTypeFromAggregateDataset} from 'sentry/views/alerts/wizard/utils';
+import {getAnomalyMarkerSeries} from 'sentry/views/detectors/utils/anomalyChart';
+import {isCrashFreeAlert} from 'sentry/views/detectors/utils/isCrashFreeAlert';
 
 function formatTooltipDate(date: moment.MomentInput, format: string): string {
   return moment(date).format(format);
