+++ static/app/utils/timeSeries/determineSeriesConfidence.spec.tsx
@@ -1,6 +1,6 @@
 import {TimeSeriesFixture} from 'sentry-fixture/timeSeries';
 
-import {determineTimeSeriesConfidence} from 'sentry/views/alerts/rules/metric/utils/determineSeriesConfidence';
+import {determineTimeSeriesConfidence} from 'sentry/utils/timeSeries/determineSeriesConfidence';
 
 describe('determineTimeSeriesConfidence', () => {
   it('equal null if no data', () => {
