+++ static/app/utils/timeSeries/determineSeriesConfidence.tsx
