+++ static/app/utils/timeSeries/determineSeriesSampleCount.tsx
