+++ static/app/views/detectors/hooks/useMetricDetectorAnomalyPeriods.tsx
@@ -10,14 +10,14 @@ import {
 } from 'sentry/views/alerts/rules/metric/types';
 import type {Anomaly} from 'sentry/views/alerts/types';
 import {AnomalyType} from 'sentry/views/alerts/types';
+import {DetectorDataset} from 'sentry/views/detectors/datasetConfig/types';
+import {isEapDataset} from 'sentry/views/detectors/datasetConfig/utils/isEapDataset';
+import {useMetricDetectorSeries} from 'sentry/views/detectors/hooks/useMetricDetectorSeries';
 import {
   EAP_HISTORICAL_TIME_PERIOD_MAP,
   HISTORICAL_TIME_PERIOD_MAP,
   HISTORICAL_TIME_PERIOD_MAP_FIVE_MINS,
-} from 'sentry/views/alerts/utils/timePeriods';
-import {DetectorDataset} from 'sentry/views/detectors/datasetConfig/types';
-import {isEapDataset} from 'sentry/views/detectors/datasetConfig/utils/isEapDataset';
-import {useMetricDetectorSeries} from 'sentry/views/detectors/hooks/useMetricDetectorSeries';
+} from 'sentry/views/detectors/utils/timePeriods';
 
 import type {IncidentPeriod} from './useIncidentMarkers';
 import {useMetricDetectorAnomalies} from './useMetricDetectorAnomalies';
