+++ static/app/views/detectors/utils/anomalyChart.tsx
