+++ static/app/views/detectors/utils/isCrashFreeAlert.tsx
