+++ static/app/views/detectors/utils/timePeriods.spec.tsx
