+++ static/app/views/explore/logs/exports/useLogsExportEstimatedRowCount.tsx
@@ -1,6 +1,6 @@
 import {defined} from 'sentry/utils/defined';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
-import {determineSeriesSampleCountAndIsSampled} from 'sentry/views/alerts/rules/metric/utils/determineSeriesSampleCount';
+import {determineSeriesSampleCountAndIsSampled} from 'sentry/utils/timeSeries/determineSeriesSampleCount';
 import {
   useProgressiveQuery,
   type RPCQueryExtras,
