+++ static/app/views/explore/logs/useLogsTimeseries.tsx
@@ -2,7 +2,7 @@ import {useCallback} from 'react';
 
 import {defined} from 'sentry/utils/defined';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
-import {determineSeriesSampleCountAndIsSampled} from 'sentry/views/alerts/rules/metric/utils/determineSeriesSampleCount';
+import {determineSeriesSampleCountAndIsSampled} from 'sentry/utils/timeSeries/determineSeriesSampleCount';
 import {
   useProgressiveQuery,
   type RPCQueryExtras,
